+++ R/modelsummary.R
@@ -20,14 +20,14 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low',
 #' @param output filename or object type (character string)
 #' * Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.
 #' * Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "modelsummary_list", "gt", "kableExtra", "huxtable", "flextable", "jupyter".
-#' * To change the default output format, type `options(modelsummary_default = "latex")`, where `latex` can be any of the valid object types listed above.
-#' * Warning: users should not supply a file name to the `output` argument if they intend to customize the table with external packages.
-#' * See the 'Details' section below for more information.
+#' * Warning: Users should not supply a file name to the `output` argument if they intend to customize the table with external packages. See the 'Details' section.
+#' * LaTeX compilation requires the `booktabs` and `siunitx` packages, but `siunitx` can be disabled or replaced with global options. See the 'Details' section.
+#' * The default output formats and table-making packages can be modified with global options. See the 'Details' section.
 #' @param fmt determines how to format numeric values
 #' * integer: the number of digits to keep after the period `format(round(x, fmt), nsmall=fmt)`
 #' * character: passed to the `sprintf` function (e.g., '%.3f' keeps 3 digits with trailing zero). See `?sprintf`
 #' * function: returns a formatted character string.
-#' * LaTeX Warning: To ensure proper formatting, all numeric entries are enclosed in the `\num{}` command by default. This command is supplied by the `siunitx` LaTeX package, so the following code must appear in the LaTeX document preamble (it is added automatically when compiling Rmarkdown documents): `\usepackage[input-symbols=()]{siunitx}` This behavior can be altered using the `math_latex` global option (see Details section).
+#' * LaTeX tables: All numeric entries are enclosed in the `siunitx` package's `\num{}` command by default. This behavior can be altered with global options. See the 'Details' section.
 #' @param stars to indicate statistical significance
 #' * FALSE (default): no significance stars.
 #' * TRUE: +=.1, *=.05, **=.01, ***=0.001
@@ -101,7 +101,10 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low',
 #' * "l": left-aligned column
 #' * "c": centered column
 #' * "r": right-aligned column
-#' * "S": math mode column. In HTML tables, numeric values are centered and wrapped in `$$` in order to be interpreted by MathJax. In LaTeX and PDF documents, numeric values are aligned on the dot and treated as math mode, using the "S" column type supplied by the `siunitx` LaTeX package. This code must appear in the LaTeX document preamble (it is added automatically when compiling Rmarkdown documents): `\usepackage[parse-numbers=false]{siunitx}` Warning: When using `siunitx` for math mode tables in LaTeX, characters like underscores in variable names will *not* be escaped automatically, and may break compilation. 
+#' * "d": dot-aligned column. Only supported for LaTeX/PDF tables produced by `kableExtra`. These commands must appear in the LaTeX preamble (they are added automatically when compiling Rmarkdown documents to PDF):
+#'   - `\usepackage{booktabs}`
+#'   - `\usepackage{siunitx}`
+#'   - `\newcolumntype{d}{S[input-open-uncertainty=,input-close-uncertainty=,table-align-text-pre=false]}`
 #' @param escape boolean TRUE escapes or substitutes LaTeX/HTML which could
 #'     prevent the file from compiling or displaying. This setting does not
 #'     affect captions or notes.
@@ -140,11 +143,16 @@ globalVariables(c('.', 'term', 'part', 'estimate', 'conf.high', 'conf.low',
 #' `options(modelsummary_get = "all")`
 #'
 #' By default, LaTeX tables enclose all numeric entries in the `\num{}` command
-#' from the siunitx package. To prevent this behavior, or to enclose
-#' numbers is math mode dollar signs instead, users can call:
+#' from the siunitx package. To prevent this behavior, or to enclose numbers
+#' in dollar signs (for LaTeX math mode), users can call:
 #'
-#' `options(modelsummary_math_latex = "nothing")`
-#' `options(modelsummary_math_latex = "dollars")`
+#' `options(modelsummary_format_numeric_latex = "plain")`
+#' `options(modelsummary_format_numeric_latex = "mathmode")`
+#'
+#' A similar option can be used to display numerical entries using MathJax in
+#' HTML tables:
+#'
+#' `options(modelsummary_format_numeric_html = "mathjax")`
 #'
 #' `output` argument:
 #'