+++ R/settings.R
@@ -4,8 +4,8 @@ settings_init <- function(settings = NULL) {
     settings_rm()
 
     default_settings <- list(
-        "math_latex" = getOption("modelsummary_math_latex", default = "siunitx"),
-        "math_html" = getOption("modelsummary_math_html", default = "minus"),
+        "format_numeric_latex" = getOption("modelsummary_format_numeric_latex", default = "siunitx"),
+        "format_numeric_html" = getOption("modelsummary_format_numeric_html", default = "minus"),
         "siunitx_scolumns" = FALSE,
         "output_default" = getOption("modelsummary_default", default = "kableExtra"),
         "stars_note" = getOption("modelsummary_stars_note", default = TRUE))