+++ tests/testthat/test-mathmode.R
@@ -15,21 +15,21 @@ mod <- list(
 ## library(modelsummary)
 
 
-test_that("S-column: supported outputs", {
-    expect_error(modelsummary(mod, output = "latex", align = "lSS"), NA)
-    expect_error(modelsummary(mod, output = "latex_tabular", align = "lSS"), NA)
-    expect_error(modelsummary(mod, output = "html", align = "lSS"), "only supported")
-    expect_error(modelsummary(mod, output = "kableExtra", align = "lSS"), "only supported")
-    expect_error(modelsummary(mod, output = "gt", align = "lSS"), regexp = "only supported")
-    expect_error(modelsummary(mod, output = "huxtable", align = "lSS"), regexp = "only supported")
-    expect_error(modelsummary(mod, output = "flextable", align = "lSS"), regexp = "only supported")
-    expect_error(modelsummary(mod, output = "markdown", align = "lSS"), regexp = "only supported")
+test_that("d-column: supported outputs", {
+    expect_error(modelsummary(mod, output = "latex", align = "ldd"), NA)
+    expect_error(modelsummary(mod, output = "latex_tabular", align = "ldd"), NA)
+    expect_error(modelsummary(mod, output = "html", align = "ldd"), "only supported")
+    expect_error(modelsummary(mod, output = "kableExtra", align = "ldd"), "only supported")
+    expect_error(modelsummary(mod, output = "gt", align = "ldd"), regexp = "only supported")
+    expect_error(modelsummary(mod, output = "huxtable", align = "ldd"), regexp = "only supported")
+    expect_error(modelsummary(mod, output = "flextable", align = "ldd"), regexp = "only supported")
+    expect_error(modelsummary(mod, output = "markdown", align = "ldd"), regexp = "only supported")
 })
 
 
 
 test_that("S-column: known output", {
-    expect_known_output(modelsummary(mod, align = "lSS", output = "latex"),
+    expect_known_output(modelsummary(mod, align = "ldd", output = "latex"),
                         file = "known_output/mathmode_1.tex",
                         print = TRUE,
                         update = TRUE)
@@ -42,17 +42,17 @@ test_that("HTML global options", {
                         file = "known_output/mathmode_2.tex",
                         print = TRUE,
                         update = TRUE)
-    options("modelsummary_math_html" = "dollars")
+    options("modelsummary_format_numeric_html" = "dollars")
     expect_known_output(modelsummary(mod, output = "html"),
                         file = "known_output/mathmode_3.tex",
                         print = TRUE,
                         update = TRUE)
-    options("modelsummary_math_html" = "anything else")
+    options("modelsummary_format_numeric_html" = "anything else")
     expect_known_output(modelsummary(mod, output = "html"),
                         file = "known_output/mathmode_4.tex",
                         print = TRUE,
                         update = TRUE)
-    options("modelsummary_math_html" = NULL)
+    options("modelsummary_format_numeric_html" = NULL)
 })
 
 
@@ -62,15 +62,15 @@ test_that("LaTeX global options", {
                         file = "known_output/mathmode_5.tex",
                         print = TRUE,
                         update = TRUE)
-    options("modelsummary_math_latex" = "dollars")
+    options("modelsummary_format_numeric_latex" = "dollars")
     expect_known_output(modelsummary(mod, output = "latex"),
                         file = "known_output/mathmode_6.tex",
                         print = TRUE,
                         update = TRUE)
-    options("modelsummary_math_latex" = "anything else")
+    options("modelsummary_format_numeric_latex" = "anything else")
     expect_known_output(modelsummary(mod, output = "latex"),
                         file = "known_output/mathmode_7.tex",
                         print = TRUE,
                         update = TRUE)
-    options("modelsummary_math_latex" = NULL)
+    options("modelsummary_format_numeric_latex" = NULL)
 })